File numbers are now ints instead of bytes.
Does not affect the C++ side of the code yet, this should break
compilation.

+++ src/main/java/im/tox/tox4j/v2/ToxCore.java
@@ -186,7 +186,7 @@ public interface ToxCore extends Closeable {
 
     void callbackFileControl(FileControlCallback callback);
 
-    byte fileSend(int friendNumber, ToxFileKind kind, long fileSize, byte[] filename) throws ToxFileSendException;
+    int fileSend(int friendNumber, ToxFileKind kind, long fileSize, byte[] filename) throws ToxFileSendException;
 
     void fileSendChunk(int friendNumber, byte fileNumber, byte[] data) throws ToxFileSendChunkException;
